+++ routes/controller/namespace-setting.go
@@ -198,7 +198,7 @@ func bindNamespaceSettingController(ctx *RouterContext) {
 			} else {
 				userList = ns.ACL.ACL
 			}
-			totalMemberCount := len(userList)
+			totalMemberCount := int64(len(userList))
 			pageInfo, err := GeneratePageInfo(r, totalMemberCount)
 			// the reason we do this is the fact that maps in go does not
 			// guarantee the order of keys when doing a range over them.
